[APM] Track usage of Gold+ features (#77630)
* adding license check

* fixing api test

* refactoring
