+++ x-pack/plugins/apm/server/routes/settings/custom_link.ts
@@ -3,9 +3,14 @@
  * or more contributor license agreements. Licensed under the Elastic License;
  * you may not use this file except in compliance with the Elastic License.
  */
+
+import Boom from 'boom';
 import * as t from 'io-ts';
 import { pick } from 'lodash';
+import { INVALID_LICENSE } from '../../../common/custom_link';
+import { ILicense } from '../../../../licensing/common/types';
 import { FILTER_OPTIONS } from '../../../common/custom_link/custom_link_filter_options';
+import { notifyFeatureUsage } from '../../feature';
 import { setupRequest } from '../../lib/helpers/setup_request';
 import { createOrUpdateCustomLink } from '../../lib/settings/custom_link/create_or_update_custom_link';
 import {
@@ -16,6 +21,10 @@ import { deleteCustomLink } from '../../lib/settings/custom_link/delete_custom_l
 import { getTransaction } from '../../lib/settings/custom_link/get_transaction';
 import { listCustomLinks } from '../../lib/settings/custom_link/list_custom_links';
 import { createRoute } from '../create_route';
+
+function isActiveGoldLicense(license: ILicense) {
+  return license.isActive && license.hasAtLeast('gold');
+}
 
 export const customLinkTransactionRoute = createRoute(() => ({
   path: '/api/apm/settings/custom_links/transaction',
@@ -37,6 +46,9 @@ export const listCustomLinksRoute = createRoute(() => ({
     query: filterOptionsRt,
   },
   handler: async ({ context, request }) => {
+    if (!isActiveGoldLicense(context.licensing.license)) {
+      throw Boom.forbidden(INVALID_LICENSE);
+    }
     const setup = await setupRequest(context, request);
     const { query } = context.params;
     // picks only the items listed in FILTER_OPTIONS
@@ -55,9 +67,17 @@ export const createCustomLinkRoute = createRoute(() => ({
     tags: ['access:apm', 'access:apm_write'],
   },
   handler: async ({ context, request }) => {
+    if (!isActiveGoldLicense(context.licensing.license)) {
+      throw Boom.forbidden(INVALID_LICENSE);
+    }
     const setup = await setupRequest(context, request);
     const customLink = context.params.body;
     const res = await createOrUpdateCustomLink({ customLink, setup });
+
+    notifyFeatureUsage({
+      licensingPlugin: context.licensing,
+      featureName: 'customLinks',
+    });
     return res;
   },
 }));
@@ -75,6 +95,9 @@ export const updateCustomLinkRoute = createRoute(() => ({
     tags: ['access:apm', 'access:apm_write'],
   },
   handler: async ({ context, request }) => {
+    if (!isActiveGoldLicense(context.licensing.license)) {
+      throw Boom.forbidden(INVALID_LICENSE);
+    }
     const setup = await setupRequest(context, request);
     const { id } = context.params.path;
     const customLink = context.params.body;
@@ -99,6 +122,9 @@ export const deleteCustomLinkRoute = createRoute(() => ({
     tags: ['access:apm', 'access:apm_write'],
   },
   handler: async ({ context, request }) => {
+    if (!isActiveGoldLicense(context.licensing.license)) {
+      throw Boom.forbidden(INVALID_LICENSE);
+    }
     const setup = await setupRequest(context, request);
     const { id } = context.params.path;
     const res = await deleteCustomLink({
